% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medsim-data.R
\docType{data}
\name{medsim}
\alias{medsim}
\title{The simulated data set ispired by the medication example}
\format{
`medsim`
a data.frame with 5 columns and 420 rows:
\describe{
\item{obs}{row number}
\item{id}{participant number}
\item{time}{time in waves from 0 to 18, with increments of 3}
\item{treat}{treatment assignment, 1= treatment; 0= control}
\item{pos}{the positive mood score}}
}
\usage{
data(medsim)
}
\description{
A simulated dataset
}
\keyword{datasets}
